Create disney_movie_club.php
My kids love Disney so we've signed up for the Disney movie club, problem is the movies will automatically be charged and shipped if you don't opt out each month.  This nagios script alerts me when a movie needs to be acknowledged so that I am not charged unless I want to be.